Update MSDK to v11.0.0 (#30)
* Bump version to 6.0.0

* Update Mimi SDK to v11

* Migrate to MimiSDK v11.0.0

* Update CHANGELOG

+++ IntegrationExample/Utilities/DebouncedSlider.swift
@@ -11,14 +11,14 @@ import Combine
 /// A slider that debounces the user input values. Default debounce interval of 100ms.
 struct DebouncedSlider<Label, ValueLabel> : View where Label : View, ValueLabel : View {
     
-    @Binding var value: Float
+    @Binding var value: Double
     @StateObject private var viewModel: DebouncedSliderViewModel
     
     private var labelView: Label
     private var minimumValueLabelView: ValueLabel
     private var maximumValueLabelView: ValueLabel
     
-    init(value: Binding<Float>, debounceInterval: Double = 0.1, @ViewBuilder label: () -> Label, @ViewBuilder minimumValueLabel: () -> ValueLabel, @ViewBuilder maximumValueLabel: () -> ValueLabel) {
+    init(value: Binding<Double>, debounceInterval: Double = 0.1, @ViewBuilder label: () -> Label, @ViewBuilder minimumValueLabel: () -> ValueLabel, @ViewBuilder maximumValueLabel: () -> ValueLabel) {
         self._value = value
         _viewModel = StateObject(wrappedValue: DebouncedSliderViewModel(value: value.wrappedValue, debounceInterval: debounceInterval))
         labelView = label()
@@ -45,10 +45,10 @@ struct DebouncedSlider<Label, ValueLabel> : View where Label : View, ValueLabel 
 extension DebouncedSlider {
     
     private class DebouncedSliderViewModel : ObservableObject {
-        @Published var debouncedValue: Float
-        @Published var value: Float
-        
-        init(value: Float, debounceInterval: Double) {
+        @Published var debouncedValue: Double
+        @Published var value: Double
+
+        init(value: Double, debounceInterval: Double) {
             self.value = value
             debouncedValue = value
             $value
